+++ RELEASE_NOTES.md
@@ -1,5 +1,6 @@
 ### 1.4.0.0-beta3 - Unreleased
 * Migration to xUnit 2.0
+* FsUnit.NUnit compiled for net45
 * FsUnitDepricated module renamed to FsUnitDeprecated https://github.com/fsprojects/FsUnit/pull/46
 * Enable substring checks https://github.com/fsprojects/FsUnit/pull/45
 * Fixed assertion message for NUnit equal constraint - https://github.com/fsprojects/FsUnit/pull/60
@@ -8,7 +9,7 @@
 * Added support of `Empty` operator for xUnit, MsTest and MbUnit - https://github.com/fsprojects/FsUnit/pull/66
 * Added support of `NaN` operator for xUnit, MsTest and MbUnit - https://github.com/fsprojects/FsUnit/pull/67
 * Added support of `haveLength` and `haveCount` operators for xUnit, MsTest and MbUnit - https://github.com/fsprojects/FsUnit/pull/68
-* Added support of `unique` operator for xUnit, MsTest and MbUnit -
+* Added support of `unique` operator for xUnit, MsTest and MbUnit - https://github.com/fsprojects/FsUnit/pull/70
 
 ### 1.3.1.0 - 26 July 2015
 * Bump NUnit version up to 2.6.4